ARM: layerscape: ppa: fix type of C variable used in inline assembly
clang warns that cr should be an unsigned long as it's used to
initialize a register argument for the mrs instruction.

Change the type from unsigned int to unsigned long to fix this.

Link: https://lore.barebox.org/20240108095959.1249068-1-a.fatoum@pengutronix.de

+++ arch/arm/mach-layerscape/ppa.c
@@ -123,7 +123,7 @@ int ls1046a_ppa_init(resource_size_t ppa_start, resource_size_t ppa_size)
 	get_builtin_firmware(ppa_ls1046a_bin, &ppa_fw, &ppa_fw_size);
 
 	if (el == 3) {
-		unsigned int cr;
+		unsigned long cr;
 
 		asm volatile("mrs %0, sctlr_el3" : "=r" (cr) : : "cc");
 		remap_range((void *)ppa_start, ppa_size, MAP_CACHED);
